% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mjsArea.R
\name{mjsArea}
\alias{mjsArea}
\title{Area charts}
\usage{
mjsArea(morrisjs, options = NULL)
}
\description{
Draw an area chart. Possible options are available 
\href{http://morrisjs.github.io/morris.js/lines.html}{here}.
}

