% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mjsBar.R
\name{mjsBar}
\alias{mjsBar}
\title{Bar charts}
\usage{
morrisjs(data) %>% 
mjsBar(options = NULL)
}
\arguments{
\item{morrisjs}{a \code{\link{morrisjs}} object,}

\item{options}{an optional list of parameters}
}
\description{
Draw a bar chart. Possible options are available 
\href{http://morrisjs.github.io/morris.js/bars.html}{here}.
}

