% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mjsDonut.R
\name{mjsDonut}
\alias{mjsDonut}
\title{Donut charts}
\usage{
morrisjs(data) %>% 
mjsArea(options = NULL)
}
\arguments{
\item{morrisjs}{a \code{\link{morrisjs}} object,}

\item{options}{an optional list of parameters}
}
\description{
Draw a donut chart. Possible options are available 
\href{http://morrisjs.github.io/morris.js/donuts.html}{here}.
}
\details{
Donut charts behave differently from the other charts: input data should
be a list of two elements, \code{labels} and \code{values}
}

