% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morrisjs.R
\name{morrisjs}
\alias{morrisjs}
\title{morris.js plot}
\usage{
morrisjs(data, width = NULL, height = NULL)
}
\arguments{
\item{data}{the data to be drawn, can be a \code{ts}, \code{mts}, \code{xts},
\code{data.frame} or \code{tbl_df} object.}

\item{width}{the width of the widget (in pixels),}

\item{height}{the height of the widget (in pixels).}
}
\description{
This function prepares the widget to be drawn.
}
\details{
In the case of a \code{data.frame} or a \code{tbl_df}, the first column must
be a object that can be interpreted as a \code{Data}, the other columns being
the data values.
}

