% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morrisjs.R
\name{morrisjsOutput}
\alias{morrisjsOutput}
\title{Widget output function for use in Shiny}
\usage{
morrisjsOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{a character, the output variable to read from}

\item{width}{the width of the widget (must be valid CSS),}

\item{height}{the height of the widget (must be valid CSS)}
}
\description{
Widget output function for use in Shiny
}

