% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mjsBar.R
\name{mjsBar}
\alias{mjsBar}
\title{Bar charts}
\usage{
mjsBar(morrisjs, options = NULL)
}
\description{
Draw a bar chart. Possible options are available 
\href{http://morrisjs.github.io/morris.js/bars.html}{here}.
}

