% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mjsLine.R
\name{mjsLine}
\alias{mjsLine}
\title{Line charts}
\usage{
mjsLine(morrisjs, options = NULL)
}
\arguments{
\item{morrisjs}{a \code{\link{morrisjs}} object,}

\item{options}{an optional list of parameters}
}
\description{
Draw a line chart. Possible options are available 
\href{http://morrisjs.github.io/morris.js/lines.html}{here}.
}

