% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morrisjs-package.r, R/morrisjs.R
\docType{package}
\name{morrisjs}
\alias{morrisjs}
\alias{morrisjs-package}
\title{morrisjs.}
\usage{
morrisjs(message, width = NULL, height = NULL)
}
\description{
morrisjs.

<Add Description>
}

