% Generated by roxygen2 (5.0.1): do not edit by hand
% Please edit documentation in R/morrisjs.R
\name{morrisjsOutput}
\alias{morrisjsOutput}
\title{Widget output function for use in Shiny}
\usage{
morrisjsOutput(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}

