% Generated by roxygen2 (5.0.1): do not edit by hand
% Please edit documentation in R/morrisjs.R
\name{renderMorrisjs}
\alias{renderMorrisjs}
\title{Widget render function for use in Shiny}
\usage{
renderMorrisjs(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

